+++ app/src/main/java/dev/pandesal/sbp/domain/repository/RecurringTransactionRepositoryInterface.kt
@@ -5,6 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface RecurringTransactionRepositoryInterface {
     fun getRecurringTransactions(): Flow<List<RecurringTransaction>>
+    fun getRecurringTransactionById(id: String): Flow<RecurringTransaction>
     suspend fun addRecurringTransaction(transaction: RecurringTransaction)
     suspend fun removeRecurringTransaction(transaction: RecurringTransaction)
 }