+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.29.0] - 2025-05-19
+### Added
+- Navigate to Recurring Transactions from Settings.
+
+
+## [0.29.0] - 2025-05-19
+### Added
+- View recurring transaction details with edit and delete actions.
+- View upcoming recurring transactions in a new screen.
 
 ## [0.28.1] - 2025-05-19
 ### Changed